fixing one

+++ src/main/java/core/basesyntax/MyLinkedList.java
@@ -9,13 +9,12 @@ public class MyLinkedList<T> implements MyLinkedListInterface<T> {
 
     @Override
     public void add(T value) {
-        final Node<T> l = last;
-        final Node<T> newNode = new Node<>(l, value, null);
-        last = newNode;
-        if (l == null) {
-            first = newNode;
+        if (last == null) {
+            first = new Node<>(null, value, null);
+            last = first;
         } else {
-            l.next = newNode;
+            last.next = new Node<>(last, value,null);
+            last = last.next;
         }
         size++;
     }
@@ -24,18 +23,16 @@ public class MyLinkedList<T> implements MyLinkedListInterface<T> {
     public void add(T value, int index) {
         if (index == size) {
             add(value);
-            return;
         } else {
             checkIndex(index);
-            Node<T> pred = node(index).prev;
-            Node<T> newNode = new Node<>(pred, value, node(index));
-            node(index).prev = newNode;
-            if (pred == null) {
-                first = newNode;
+            if (index == 0) {
+                first.prev = new Node<>(null, value, first);
+                first = first.prev;
             } else {
-                pred.next = newNode;
+                Node<T> newNode = new Node<>(nodeByIndex(index).prev, value, nodeByIndex(index));
+                nodeByIndex(index).prev.next = newNode;
+                nodeByIndex(index).prev = newNode;
             }
-            pred = newNode;
             size++;
         }
     }
@@ -50,41 +47,36 @@ public class MyLinkedList<T> implements MyLinkedListInterface<T> {
     @Override
     public T get(int index) {
         checkIndex(index);
-        return node(index).item;
+        return nodeByIndex(index).item;
     }
 
     @Override
     public T set(T value, int index) {
         checkIndex(index);
-        Node<T> x = node(index);
-        T oldVal = x.item;
-        x.item = value;
+        Node<T> node = nodeByIndex(index);
+        T oldVal = node.item;
+        node.item = value;
         return oldVal;
     }
 
     @Override
     public T remove(int index) {
         checkIndex(index);
-        T element = node(index).item;
-        unlink(node(index));
+        T element = nodeByIndex(index).item;
+        unlink(nodeByIndex(index));
         return element;
     }
 
     @Override
     public boolean remove(T object) {
-        if (object == null) {
-            for (Node<T> x = first; x != null; x = x.next) {
-                if (x.item == null) {
-                    unlink(x);
-                    return true;
-                }
+        for (Node<T> x = first; x != null; x = x.next) {
+            if (object == null && x.item == null) {
+                unlink(x);
+                return true;
             }
-        } else {
-            for (Node<T> x = first; x != null; x = x.next) {
-                if (object.equals(x.item)) {
-                    unlink(x);
-                    return true;
-                }
+            if (object != null && object.equals(x.item)) {
+                unlink(x);
+                return true;
             }
         }
         return false;
@@ -118,7 +110,7 @@ public class MyLinkedList<T> implements MyLinkedListInterface<T> {
         }
     }
 
-    Node<T> node(int index) {
+    Node<T> nodeByIndex(int index) {
         if (index < (size >> 1)) {
             Node<T> x = first;
             for (int i = 0; i < index; i++) {
@@ -137,21 +129,18 @@ public class MyLinkedList<T> implements MyLinkedListInterface<T> {
     private void unlink(Node<T> x) {
         final Node<T> next = x.next;
         final Node<T> prev = x.prev;
-
         if (prev == null) {
-            first = next;
+            first = x.next;
         } else {
             prev.next = next;
             x.prev = null;
         }
-
         if (next == null) {
             last = prev;
         } else {
             next.prev = prev;
             x.next = null;
         }
-
         x.item = null;
         size--;
     }
